requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
sqlparse==0.4.3
pytz==2021.1
fabric3